+++ src/formula/CAPSYS2/spectrum.py
