+++ src/formula/CAPSYS2/wall_polygon.py
